Added automatic logging based on zone.
Added CanLogInNewZone() function to determine whether or not combat should
be logging in the new zone.
Added SetLoggingStatus(status) function to enable or disabled combat
logging based on the return value of CanLogInNewZone().
Added ZoneChanged() function to call the previous two functions and handle
automatic logging.
Call ZoneChanged() when events are handled and the AddOn needs to be
updated.
Added new debug variable and display debugging text when it's enabled.