Suppress extra "logging stopped" messages.
Added a check to see if combat is currently being logged to decide if a
message should display telling player that combat has stopped being
logged.
Fixed the way CanLogInNewZone() returns it's true/false values to fix a
bug where combat would not be logged when more than two zones are in the
list of logging zones.